# swagger.yml
openapi: 3.0.1
info:
  title: MAN Trucks
  description: 'Esta é a definição da API para o code challenge'
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: https://localhost/v1
- url: http://localhost/v1
tags:
- name: truck
  description: Powering you Business on the road
  externalDocs:
    description: Find out more
    url: http://swagger.io
paths:
  /trucks:
    get:
      tags:
      - truck
      summary: List all trucks
      operationId: listTruck
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
    put:
      tags:
      - truck
      summary: Update an existing truck
      operationId: updateTruck
      requestBody:
        description: Truck object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Truck'
          application/xml:
            schema:
              $ref: '#/components/schemas/Truck'
        required: true
      responses:
        200:
          description: Truck updated successfully
          content: {}
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Truck not found
          content: {}
        405:
          description: Validation exception
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
      - truck
      summary: Adds a new truck
      operationId: addTruck
      requestBody:
        description: Truck object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Truck'
          application/xml:
            schema:
              $ref: '#/components/schemas/Truck'
        required: true
      responses:
        201:
          description: Truck added successfully
          content: {}
        405:
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /trucks/{truckId}:
    get:
      tags:
      - truck
      summary: Find truck by ID
      description: Returns a single truck
      operationId: getTruckById
      parameters:
      - name: truckId
        in: path
        description: ID of truck to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Truck'
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Truck not found
          content: {}
    post:
      tags:
      - truck
      summary: Updates a truck with form data
      operationId: updateTruckWithForm
      parameters:
      - name: truckId
        in: path
        description: ID of truck that needs to be updated
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Truck object that needs to be updated
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Truck'
        required: true
      responses:
        200:
          description: Truck updated successfully
          content: {}
        400:
          description: Invalid ID supplied
          content: {}
        405:
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
      - truck
      summary: Deletes a truck
      operationId: deleteTruck
      parameters:
      - name: truckId
        in: path
        description: Truck id to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: Truck deleted successfully
          content: {}
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Truck not found
          content: {}
components:
  schemas:
    Truck:
      required:
      - colors
      - engineHP
      - engineVolume
      - name
      - segment
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: TGX 18.44
        engineHP:
          type: integer
          example: 440
        engineVolume:
          type: integer
          example: 10518
        fuel:
          type: string
          example: diesel
        segments:
          type: array
          xml:
            name: segment
            wrapped: true
          items:
            type: string
            description: truck segment
            enum:
            - Long Haul
            - Construction
            - Firedepartment
            - Distribution (Food)
            - Wastedisposal
          example:
          - Construction
          - Firedepartment
        colors:
          type: array
          xml:
            name: color
            wrapped: true
          items:
            type: string
          example:
          - red
          - white
          - blue
      xml:
        name: Truck
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string